More polished initial camera API + improve type consistency (#639)
* Cleanup

* Re-generate docs

* Add new files

* Simplify, add test to confirm that Python and TS initial states match

* Simplify

* Consistent type annotations, adjust initial message behavior

* Fix comments

* Add future import

+++ src/viser/_scene_handles.py
@@ -133,7 +133,7 @@ class SceneNodeHandle(AssignablePropsBase[_SceneNodeHandleState]):
         return out
 
     @property
-    def wxyz(self) -> npt.NDArray[np.float32]:
+    def wxyz(self) -> npt.NDArray[np.float64]:
         """Orientation of the scene node. This is the quaternion representation of the R
         in `p_parent = [R | t] p_local`. Synchronized to clients automatically when assigned.
         """
@@ -153,7 +153,7 @@ class SceneNodeHandle(AssignablePropsBase[_SceneNodeHandleState]):
         )
 
     @property
-    def position(self) -> npt.NDArray[np.float32]:
+    def position(self) -> npt.NDArray[np.float64]:
         """Position of the scene node. This is equivalent to the t in
         `p_parent = [R | t] p_local`. Synchronized to clients automatically when assigned.
         """
@@ -661,7 +661,7 @@ class MeshSkinnedBoneHandle:
     _impl: BoneState
 
     @property
-    def wxyz(self) -> npt.NDArray[np.float32]:
+    def wxyz(self) -> npt.NDArray[np.float64]:
         """Orientation of the bone. This is the quaternion representation of the R
         in `p_parent = [R | t] p_local`. Synchronized to clients automatically when assigned.
         """
@@ -683,7 +683,7 @@ class MeshSkinnedBoneHandle:
         )
 
     @property
-    def position(self) -> npt.NDArray[np.float32]:
+    def position(self) -> npt.NDArray[np.float64]:
         """Position of the bone. This is equivalent to the t in
         `p_parent = [R | t] p_local`. Synchronized to clients automatically when assigned.
         """
